+++ src/components/progress.tsx
@@ -1,23 +1,36 @@
 import { forwardRef, useContext } from 'react';
 import styled from 'styled-components';
 import { DeckContext } from './deck/deck';
-import { position } from 'styled-system';
+import { position, PositionProps } from 'styled-system';
 
-export type CircleProps = { size: number; color: string; active: boolean };
+const DEFAULT_PROGRESS_CIRCLE_SIZE = 10;
+export const PROGRESS_CIRCLE_BORDER_WIDTH = 1;
+const PROGRESS_CIRCLE_MARGIN = DEFAULT_PROGRESS_CIRCLE_SIZE / 3;
+export const DEFAULT_PROGRESS_CIRCLE_WIDTH_INCLUDING_MARGIN =
+  DEFAULT_PROGRESS_CIRCLE_SIZE +
+  (PROGRESS_CIRCLE_BORDER_WIDTH + PROGRESS_CIRCLE_MARGIN) * 2;
+
+export type CircleProps = {
+  size: number;
+  margin: number;
+  color: string;
+  active: boolean;
+};
 export const Circle = styled.div<CircleProps>`
   width: ${({ size }) => size}px;
   height: ${({ size }) => size}px;
-  display: inline-block;
-  border: 1px solid ${({ color }) => color};
+  border: ${PROGRESS_CIRCLE_BORDER_WIDTH}px solid ${({ color }) => color};
   background: ${({ color, active }) => (active ? color : 'transparent')};
-  margin: ${({ size }) => size / 3}px;
+  margin: ${({ margin }) => margin}px;
   border-radius: 50%;
   pointer-events: all;
   cursor: pointer;
 `;
 
-const Container = styled.div`
+export const ProgressContainer = styled.div<PositionProps>`
   ${position}
+  display: flex;
+  flex-wrap: wrap;
   @media print {
     display: none;
   }
@@ -29,10 +42,14 @@ export type ProgressProps = {
 };
 
 const Progress = forwardRef<HTMLDivElement, ProgressProps>(
-  ({ color = '#fff', size = 10, ...props }, ref) => {
+  ({ color = '#fff', size = DEFAULT_PROGRESS_CIRCLE_SIZE, ...props }, ref) => {
     const { slideCount, skipTo, activeView } = useContext(DeckContext);
     return (
-      <Container ref={ref} className="spectacle-progress-indicator" {...props}>
+      <ProgressContainer
+        ref={ref}
+        className="spectacle-progress-indicator"
+        {...props}
+      >
         {Array(slideCount)
           .fill(0)
           .map((_, idx) => (
@@ -41,6 +58,7 @@ const Progress = forwardRef<HTMLDivElement, ProgressProps>(
               color={color}
               active={activeView.slideIndex === idx}
               size={size}
+              margin={PROGRESS_CIRCLE_MARGIN}
               onClick={() =>
                 skipTo({
                   slideIndex: idx,
@@ -50,7 +68,7 @@ const Progress = forwardRef<HTMLDivElement, ProgressProps>(
               data-testid="Progress Circle"
             />
           ))}
-      </Container>
+      </ProgressContainer>
     );
   }
 );