+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Persist deck template between slides in default and presenter modes [#1106](https://github.com/FormidableLabs/spectacle/issues/1106)
 - Upgrade dependencies to be React 18-friendly, removing Enzyme in favor of RTL via [#1119](https://github.com/FormidableLabs/spectacle/pull/1119)
 - Fix only the first page showing in print and export modes in Firefox [#1077](https://github.com/FormidableLabs/spectacle/issues/1077)
+- Add AnimatedProgress component [#1105](https://github.com/FormidableLabs/spectacle/issues/1105)
 
 ## 9.1.1